# hms-project/docs/ambulance-management-openapi.yaml
openapi: 3.0.3
info:
  title: HMS Ambulance Management API
  description: >-
    This OpenAPI specification documents all ambulance management related
    endpoints, including ambulance inventory management, dispatch system,
    real-time location tracking, driver management, service history,
    and billing integration.
  version: 1.0.0
servers:
  - url: https://hms.example.com/api
    description: Production server
  - url: http://localhost:3000/api
    description: Local dev server
...
paths:
...
  /ambulances/drivers:
    get:
      tags: [Driver Management]
      summary: Get ambulance drivers with optional filtering
      parameters:
        - in: query
          name: ambulanceId
          schema:
            type: string
        - in: query
          name: available
          schema:
            type: boolean
        - in: query
          name: id
          schema:
            type: string
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 10 }
      responses:
        '200':
          description: List of ambulance drivers with pagination metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverListResponse'
    post:
      tags: [Driver Management]
      summary: Assign a driver to an ambulance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DriverInput'
      responses:
        '201':
          description: Driver assigned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Driver'
    put:
      tags: [Driver Management]
      summary: Update driver assignment details
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/DriverInput'
                - type: object
                  properties:
                    id:
                      type: string
                  required: [id]
      responses:
        '200':
          description: Driver assignment updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Driver'
    delete:
      tags: [Driver Management]
      summary: Remove a driver assignment
      parameters:
        - in: query
          name: id
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Driver assignment removed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
  /ambulances/history:
    get:
      tags: [Service History]
      summary: Get ambulance service history
      parameters:
        - in: query
          name: ambulanceId
          schema:
            type: string
          required: true
        - in: query
          name: type
          schema:
            type: string
            enum: [MAINTENANCE, DISPATCH, ALL]
            default: ALL
        - in: query
          name: startDate
          schema:
            type: string
            format: date
        - in: query
          name: endDate
          schema:
            type: string
            format: date
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 10 }
      responses:
        '200':
          description: Ambulance service history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryResponse'
    post:
      tags: [Service History]
      summary: Add a new maintenance record to ambulance history
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaintenanceInput'
      responses:
        '201':
          description: Maintenance record created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Maintenance'
...
components:
  schemas:
    PagedResponse:
      type: object
      properties:
        data:
          type: array
          items: { type: object }
        meta:
          type: object
          properties:
            page: { type: integer }
            limit: { type: integer }
            total: { type: integer }
            totalPages: { type: integer }
...
    DispatchInput:
      type: object
      required: [pickupLocation, dropLocation, purpose, requestedBy]
      properties:
        pickupLocation:
          type: string
        dropLocation:
          type: string
        purpose:
          type: string
        priority:
          type: string
          enum: [LOW, NORMAL, HIGH, EMERGENCY]
          default: NORMAL
        patientId:
          type: string
        requestedBy:
          type: string
        notes:
          type: string
    Dispatch:
      allOf:
        - $ref: '#/components/schemas/DispatchInput'
        - type: object
          properties:
            id:
              type: string
            ambulanceId:
              type: string
            driverId:
              type: string
            status:
              type: string
              enum: [REQUESTED, ASSIGNED, DISPATCHED, ARRIVED, COMPLETED, CANCELLED]
            requestedAt:
              type: string
              format: date-time
            dispatchedAt:
              type: string
              format: date-time
            arrivedAt:
              type: string
              format: date-time
            completedAt:
              type: string
              format: date-time
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
...
    DriverListResponse:
      allOf:
        - $ref: '#/components/schemas/PagedResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Driver'
    DriverInput:
      type: object
      required: [ambulanceId, driverId, licenseNumber]
      properties:
        ambulanceId:
          type: string
        driverId:
          type: string
        licenseNumber:
          type: string
        licenseExpiry:
          type: string
          format: date
        isAvailable:
          type: boolean
          default: true
    Driver:
      allOf:
        - $ref: '#/components/schemas/DriverInput'
        - type: object
          properties:
            id:
              type: string
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    HistoryResponse:
      allOf:
        - $ref: '#/components/schemas/PagedResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                oneOf:
                  - $ref: '#/components/schemas/Maintenance'
                  - $ref: '#/components/schemas/Dispatch'
                  - $ref: '#/components/schemas/HistoryItem'
    HistoryItem:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [MAINTENANCE, DISPATCH]
        date:
          type: string
          format: date-time
        details:
          type: object
    MaintenanceInput:
      type: object
      required: [ambulanceId, maintenanceType, description, performedBy, performedAt]
      properties:
        ambulanceId:
          type: string
        maintenanceType:
          type: string
          enum: [ROUTINE, REPAIR, INSPECTION, EMERGENCY, UPGRADE, OTHER]
        description:
          type: string
        performedBy:
          type: string
        performedAt:
          type: string
          format: date-time
        cost:
          type: number
          format: double
        odometer:
          type: integer
        nextMaintenanceDue:
          type: string
          format: date
        notes:
          type: string
    Maintenance:
      allOf:
        - $ref: '#/components/schemas/MaintenanceInput'
        - type: object
          properties:
            id:
              type: string
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
...